Support to run by command line. Refactor code
