+++ build.gradle
@@ -1,12 +1,13 @@
+plugins {
+    id "io.qameta.allure" version "2.5"
+}
 group 'org.fundacionjala.pivotal'
 version '1.0'
 
 // Apply the java plugin to add support for Java
 apply plugin: 'java'
 apply plugin: 'idea'
-apply plugin: 'eclipse'
 apply plugin: 'checkstyle'
-apply plugin: 'pmd'
 apply plugin: 'findbugs'
 
 sourceCompatibility = 1.8
@@ -20,19 +21,24 @@ allprojects{
 // In this section you declare the dependencies for your production and test code
 dependencies {
     compile group: 'org.seleniumhq.selenium', name: 'selenium-java', version: '3.14.0'
-    compile group: 'io.rest-assured', name: 'rest-assured', version: '3.1.1'
+    compile group: 'io.rest-assured', name: 'rest-assured', version: '3.2.0'
     compile group: 'org.apache.logging.log4j', name: 'log4j-api', version: '2.11.0'
     compile group: 'org.apache.logging.log4j', name: 'log4j-core', version: '2.11.0'
 
     testCompile group: 'io.cucumber', name: 'cucumber-java', version: '4.0.0'
     testCompile group: 'io.cucumber', name: 'cucumber-testng', version: '4.0.0'
     testCompile group: 'io.cucumber', name: 'cucumber-picocontainer', version: '4.0.0'
-    testCompile 'io.rest-assured:rest-assured:3.2.0'
+    testCompile group: 'io.qameta.allure', name: 'allure-cucumber2-jvm', version: '2.7.0'
+}
+
+allure {
+    version = '2.7.0'
+    autoconfigure = true
 }
 
 // Checkstyle configuration
 checkstyle {
-    toolVersion = '8.12'
+    toolVersion = '8.14'
 }
 
 // FindBugs configuration
@@ -46,14 +52,20 @@ tasks.withType(FindBugs) {
     }
 }
 
-// PMD configuration
-pmd {
-    ruleSets = ["java-basic", "java-braces", "java-unnecessary", "java-unusedcode", "java-strictexception"]
-    toolVersion = '6.0.0'
-}
-tasks.withType(Pmd) {
-    reports {
-        xml.enabled = false
-        html.enabled = true
+task executeFeatures(type: Test) {
+    doFirst {
+        useTestNG()
+        options {
+            useDefaultListeners = false
+            jvmArgs = ['-Dcucumber.options= ' +
+                               '--glue org.fundacionjala.pivotal.cucumber ' +
+                               '--plugin pretty ' +
+                               '--plugin io.qameta.allure.cucumber2jvm.AllureCucumber2Jvm']
+        }
+        testLogging.showStandardStreams = true
+        scanForTestClasses = false
+        ignoreFailures = false
+        reports.html.enabled = false
     }
+    finalizedBy 'allureReport'
 }
