--- config.properties
@@ -1,12 +1,0 @@
-#Enviroment options
-browser=firefox
-url=https://www.pivotaltracker.com/
-apiURL=https://www.pivotaltracker.com/services/v5
-apiToken=8c5a5a6e8bf0cebf4e463b9c3d16bf92
-
-#Available users
-user1=hapsneeze
-password1=test12345
-
-user2=adrinrojas1
-password2=test12345
