+++ gradle.properties
@@ -1,20 +1,12 @@
-#Pivotal Properties
-urlLogin=https://www.pivotaltracker.com/signin
-user=
-password=
-APIToken=
+#Environment options
+browser=firefox
+url=https://www.pivotaltracker.com/
+apiURL=https://www.pivotaltracker.com/services/v5
+apiToken=8c5a5a6e8bf0cebf4e463b9c3d16bf92
 
-dockerURL=
-browser=
+#Available users
+user1=hapsneeze
+password1=test12345
 
-#Remote Browser SauceLabs/Browserstack properties
-remoteUserName=
-remoteAccessKey=
-remoteBrowserName=
-remotePlatform=
-remoteBrowserVersion=
-remoteResolution=
-#Browserstack property
-remoteOS=
-#Explicit Wait
-explicitWait =
+user2=adrinrojas1
+password2=test12345
