+++ src/main/java/org/fundacionjala/core/util/Environment.java
@@ -31,7 +31,7 @@ public final class Environment {
      */
     private void readFile() {
         properties = new Properties();
-        try (InputStream input = new FileInputStream("config.properties")) {
+        try (InputStream input = new FileInputStream("gradle.properties")) {
             properties.load(input);
             if (properties.getProperty("user1").isEmpty() || properties.getProperty("password1").isEmpty()) {
                 LOGGER.error("There aren't the request data");
